ingest/requirements.txt
chromadb>=0.5.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
sentence-transformers>=2.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pypdf>=4.0.0
html2text>=2024.2.0
pyyaml>=6.0
